feat(image): image refactor (#1487)
